release 3.1.87 source code

+++ core/src/main/java/com/huaweicloud/sdk/core/json/StrictFloatDeserializer.java
@@ -34,6 +34,6 @@ public class StrictFloatDeserializer extends JsonDeserializer<Float> {
 
     @Override
     public Float deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
-        return Float.valueOf(jsonParser.getFloatValue());
+        return jsonParser.getFloatValue();
     }
 }
